# design/apiPrototype/ActiveEdge Interview.postman_collection.json-APIBluePrint.apib-OpenApi3Yaml.yaml
openapi: 3.0.0
info:
  title: ActiveEdge Interview
  description: ''
  contact: {}
  version: '1.0'
servers:
- url: http://localhost:8080/api
  variables: {}
paths:
  /stocks:
    get:
      tags:
      - Misc
      summary: Get Stocks
      description: ''
      operationId: GetStocks
      parameters: []
      responses:
        '200':
          description: OK
          headers:
            Vary:
              content:
                text/plain:
                  schema:
                    type: string
                    example: Origin
                  example: Origin
            X-Content-Type-Options:
              content:
                text/plain:
                  schema:
                    type: string
                    example: nosniff
                  example: nosniff
            X-XSS-Protection:
              content:
                text/plain:
                  schema:
                    type: string
                    example: 1; mode=block
                  example: 1; mode=block
            Cache-Control:
              content:
                text/plain:
                  schema:
                    type: string
                    example: no-cache, no-store, max-age=0, must-revalidate
                  example: no-cache, no-store, max-age=0, must-revalidate
            Pragma:
              content:
                text/plain:
                  schema:
                    type: string
                    example: no-cache
                  example: no-cache
            Expires:
              content:
                text/plain:
                  schema:
                    type: string
                    example: '0'
                  example: '0'
            X-Frame-Options:
              content:
                text/plain:
                  schema:
                    type: string
                    example: DENY
                  example: DENY
            Content-Length:
              content:
                text/plain:
                  schema:
                    type: string
                    example: '152'
                  example: '152'
            Date:
              content:
                text/plain:
                  schema:
                    type: string
                    example: Sat, 27 Nov 2021 12:37:47 GMT
                  example: Sat, 27 Nov 2021 12:37:47 GMT
            Keep-Alive:
              content:
                text/plain:
                  schema:
                    type: string
                    example: timeout=60
                  example: timeout=60
            Connection:
              content:
                text/plain:
                  schema:
                    type: string
                    example: keep-alive
                  example: keep-alive
          content:
            application/json;charset=UTF-8:
              schema:
                $ref: '#/components/schemas/GetStocks'
              example:
                path: /api/stocks/
                timestamp: 2021-11-27 13:37:47
                response_state: true
                status: 200
                data:
                - name: Stock
                  amount:
                    currentPrice: '30.00'
      deprecated: false
    post:
      tags:
      - Misc
      summary: Create Stock
      description: ''
      operationId: CreateStock
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStockRequest'
            example:
              name: Stock
              amount:
                currentPrice: 30
        required: true
      responses:
        '400':
          description: Bad Request
          headers:
            Vary:
              content:
                text/plain:
                  schema:
                    type: string
                    example: Origin
                  example: Origin
            X-Content-Type-Options:
              content:
                text/plain:
                  schema:
                    type: string
                    example: nosniff
                  example: nosniff
            X-XSS-Protection:
              content:
                text/plain:
                  schema:
                    type: string
                    example: 1; mode=block
                  example: 1; mode=block
            Cache-Control:
              content:
                text/plain:
                  schema:
                    type: string
                    example: no-cache, no-store, max-age=0, must-revalidate
                  example: no-cache, no-store, max-age=0, must-revalidate
            Pragma:
              content:
                text/plain:
                  schema:
                    type: string
                    example: no-cache
                  example: no-cache
            Expires:
              content:
                text/plain:
                  schema:
                    type: string
                    example: '0'
                  example: '0'
            X-Frame-Options:
              content:
                text/plain:
                  schema:
                    type: string
                    example: DENY
                  example: DENY
            Content-Length:
              content:
                text/plain:
                  schema:
                    type: string
                    example: '190'
                  example: '190'
            Date:
              content:
                text/plain:
                  schema:
                    type: string
                    example: Sat, 27 Nov 2021 12:33:20 GMT
                  example: Sat, 27 Nov 2021 12:33:20 GMT
            Connection:
              content:
                text/plain:
                  schema:
                    type: string
                    example: close
                  example: close
          content:
            application/json;charset=UTF-8:
              schema:
                $ref: '#/components/schemas/ValidationErr1'
              example:
                path: /api/stocks
                timestamp: 2021-11-27 13:33:20
                response_state: false
                status: 400
                data:
                  message: Validation failed
                  error: Bad Request
                  errors:
                    name: Enter stock name
        '201':
          description: Created
          headers:
            Vary:
              content:
                text/plain:
                  schema:
                    type: string
                    example: Origin
                  example: Origin
            X-Content-Type-Options:
              content:
                text/plain:
                  schema:
                    type: string
                    example: nosniff
                  example: nosniff
            X-XSS-Protection:
              content:
                text/plain:
                  schema:
                    type: string
                    example: 1; mode=block
                  example: 1; mode=block
            Cache-Control:
              content:
                text/plain:
                  schema:
                    type: string
                    example: no-cache, no-store, max-age=0, must-revalidate
                  example: no-cache, no-store, max-age=0, must-revalidate
            Pragma:
              content:
                text/plain:
                  schema:
                    type: string
                    example: no-cache
                  example: no-cache
            Expires:
              content:
                text/plain:
                  schema:
                    type: string
                    example: '0'
                  example: '0'
            X-Frame-Options:
              content:
                text/plain:
                  schema:
                    type: string
                    example: DENY
                  example: DENY
            Content-Length:
              content:
                text/plain:
                  schema:
                    type: string
                    example: '108'
                  example: '108'
            Date:
              content:
                text/plain:
                  schema:
                    type: string
                    example: Sat, 27 Nov 2021 12:33:59 GMT
                  example: Sat, 27 Nov 2021 12:33:59 GMT
            Keep-Alive:
              content:
                text/plain:
                  schema:
                    type: string
                    example: timeout=60
                  example: timeout=60
            Connection:
              content:
                text/plain:
                  schema:
                    type: string
                    example: keep-alive
                  example: keep-alive
          content:
            application/json;charset=UTF-8:
              schema:
                $ref: '#/components/schemas/CreateStock'
              example:
                path: /api/stocks
                timestamp: 2021-11-27 13:33:59
                response_state: true
                status: 201
                data: Created
      deprecated: false
  /{stockId}:
    get:
      tags:
      - Misc
      summary: Get Stock
      description: ''
      operationId: GetStock
      parameters:
      - name: stockId
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: number
          format: double
          example: 1
      responses:
        '404':
          description: Not Found
          headers:
            Vary:
              content:
                text/plain:
                  schema:
                    type: string
                    example: Origin
                  example: Origin
            X-Content-Type-Options:
              content:
                text/plain:
                  schema:
                    type: string
                    example: nosniff
                  example: nosniff
            X-XSS-Protection:
              content:
                text/plain:
                  schema:
                    type: string
                    example: 1; mode=block
                  example: 1; mode=block
            Cache-Control:
              content:
                text/plain:
                  schema:
                    type: string
                    example: no-cache, no-store, max-age=0, must-revalidate
                  example: no-cache, no-store, max-age=0, must-revalidate
            Pragma:
              content:
                text/plain:
                  schema:
                    type: string
                    example: no-cache
                  example: no-cache
            Expires:
              content:
                text/plain:
                  schema:
                    type: string
                    example: '0'
                  example: '0'
            X-Frame-Options:
              content:
                text/plain:
                  schema:
                    type: string
                    example: DENY
                  example: DENY
            Content-Length:
              content:
                text/plain:
                  schema:
                    type: string
                    example: '151'
                  example: '151'
            Date:
              content:
                text/plain:
                  schema:
                    type: string
                    example: Sat, 27 Nov 2021 11:43:20 GMT
                  example: Sat, 27 Nov 2021 11:43:20 GMT
            Keep-Alive:
              content:
                text/plain:
                  schema:
                    type: string
                    example: timeout=60
                  example: timeout=60
            Connection:
              content:
                text/plain:
                  schema:
                    type: string
                    example: keep-alive
                  example: keep-alive
          content:
            application/json;charset=UTF-8:
              schema:
                $ref: '#/components/schemas/StockNotFoundErr1'
              example:
                path: /api/stocks/1
                timestamp: 2021-11-27 12:43:20
                response_state: false
                status: 404
                data:
                  message: Stock not found
                  error: Not Found
        '200':
          description: OK
          headers:
            Vary:
              content:
                text/plain:
                  schema:
                    type: string
                    example: Origin
                  example: Origin
            X-Content-Type-Options:
              content:
                text/plain:
                  schema:
                    type: string
                    example: nosniff
                  example: nosniff
            X-XSS-Protection:
              content:
                text/plain:
                  schema:
                    type: string
                    example: 1; mode=block
                  example: 1; mode=block
            Cache-Control:
              content:
                text/plain:
                  schema:
                    type: string
                    example: no-cache, no-store, max-age=0, must-revalidate
                  example: no-cache, no-store, max-age=0, must-revalidate
            Pragma:
              content:
                text/plain:
                  schema:
                    type: string
                    example: no-cache
                  example: no-cache
            Expires:
              content:
                text/plain:
                  schema:
                    type: string
                    example: '0'
                  example: '0'
            X-Frame-Options:
              content:
                text/plain:
                  schema:
                    type: string
                    example: DENY
                  example: DENY
            Content-Length:
              content:
                text/plain:
                  schema:
                    type: string
                    example: '204'
                  example: '204'
            Date:
              content:
                text/plain:
                  schema:
                    type: string
                    example: Sat, 27 Nov 2021 12:36:20 GMT
                  example: Sat, 27 Nov 2021 12:36:20 GMT
            Keep-Alive:
              content:
                text/plain:
                  schema:
                    type: string
                    example: timeout=60
                  example: timeout=60
            Connection:
              content:
                text/plain:
                  schema:
                    type: string
                    example: keep-alive
                  example: keep-alive
          content:
            application/json;charset=UTF-8:
              schema:
                $ref: '#/components/schemas/GetStock'
              example:
                path: /api/stocks/1
                timestamp: 2021-11-27 13:36:20
                response_state: true
                status: 200
                data:
                  name: Stock
                  amount:
                    currentPrice: '30.00'
                  stockId: 1
                  creationDate: 2021-11-27T01:33:59.181
      deprecated: false
    put:
      tags:
      - Misc
      summary: Update Stock
      description: ''
      operationId: UpdateStock
      parameters:
      - name: stockId
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: number
          format: double
          example: 1
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStockRequest'
            example:
              name: Stock
              amount:
                currentPrice: 30
        required: true
      responses:
        '200':
          description: OK
          headers:
            Vary:
              content:
                text/plain:
                  schema:
                    type: string
                    example: Origin
                  example: Origin
            X-Content-Type-Options:
              content:
                text/plain:
                  schema:
                    type: string
                    example: nosniff
                  example: nosniff
            X-XSS-Protection:
              content:
                text/plain:
                  schema:
                    type: string
                    example: 1; mode=block
                  example: 1; mode=block
            Cache-Control:
              content:
                text/plain:
                  schema:
                    type: string
                    example: no-cache, no-store, max-age=0, must-revalidate
                  example: no-cache, no-store, max-age=0, must-revalidate
            Pragma:
              content:
                text/plain:
                  schema:
                    type: string
                    example: no-cache
                  example: no-cache
            Expires:
              content:
                text/plain:
                  schema:
                    type: string
                    example: '0'
                  example: '0'
            X-Frame-Options:
              content:
                text/plain:
                  schema:
                    type: string
                    example: DENY
                  example: DENY
            Content-Length:
              content:
                text/plain:
                  schema:
                    type: string
                    example: '110'
                  example: '110'
            Date:
              content:
                text/plain:
                  schema:
                    type: string
                    example: Sat, 27 Nov 2021 12:36:48 GMT
                  example: Sat, 27 Nov 2021 12:36:48 GMT
            Keep-Alive:
              content:
                text/plain:
                  schema:
                    type: string
                    example: timeout=60
                  example: timeout=60
            Connection:
              content:
                text/plain:
                  schema:
                    type: string
                    example: keep-alive
                  example: keep-alive
          content:
            application/json;charset=UTF-8:
              schema:
                $ref: '#/components/schemas/UpdateStock'
              example:
                path: /api/stocks/1
                timestamp: 2021-11-27 13:36:48
                response_state: true
                status: 200
                data: Updated
        '404':
          description: Not Found
          headers:
            Vary:
              content:
                text/plain:
                  schema:
                    type: string
                    example: Origin
                  example: Origin
            X-Content-Type-Options:
              content:
                text/plain:
                  schema:
                    type: string
                    example: nosniff
                  example: nosniff
            X-XSS-Protection:
              content:
                text/plain:
                  schema:
                    type: string
                    example: 1; mode=block
                  example: 1; mode=block
            Cache-Control:
              content:
                text/plain:
                  schema:
                    type: string
                    example: no-cache, no-store, max-age=0, must-revalidate
                  example: no-cache, no-store, max-age=0, must-revalidate
            Pragma:
              content:
                text/plain:
                  schema:
                    type: string
                    example: no-cache
                  example: no-cache
            Expires:
              content:
                text/plain:
                  schema:
                    type: string
                    example: '0'
                  example: '0'
            X-Frame-Options:
              content:
                text/plain:
                  schema:
                    type: string
                    example: DENY
                  example: DENY
            Content-Length:
              content:
                text/plain:
                  schema:
                    type: string
                    example: '151'
                  example: '151'
            Date:
              content:
                text/plain:
                  schema:
                    type: string
                    example: Sat, 27 Nov 2021 12:37:25 GMT
                  example: Sat, 27 Nov 2021 12:37:25 GMT
            Keep-Alive:
              content:
                text/plain:
                  schema:
                    type: string
                    example: timeout=60
                  example: timeout=60
            Connection:
              content:
                text/plain:
                  schema:
                    type: string
                    example: keep-alive
                  example: keep-alive
          content:
            application/json;charset=UTF-8:
              schema:
                $ref: '#/components/schemas/StockNotFoundErr1'
              example:
                path: /api/stocks/2
                timestamp: 2021-11-27 13:37:25
                response_state: false
                status: 404
                data:
                  message: Stock not found
                  error: Not Found
      deprecated: false
    delete:
      tags:
      - Misc
      summary: Delete Stock
      description: ''
      operationId: DeleteStock
      parameters:
      - name: stockId
        in: path
        description: ''
        required: true
        style: simple
        schema:
          type: number
          format: double
          example: 1
      responses:
        '200':
          description: OK
          headers:
            Vary:
              content:
                text/plain:
                  schema:
                    type: string
                    example: Origin
                  example: Origin
            X-Content-Type-Options:
              content:
                text/plain:
                  schema:
                    type: string
                    example: nosniff
                  example: nosniff
            X-XSS-Protection:
              content:
                text/plain:
                  schema:
                    type: string
                    example: 1; mode=block
                  example: 1; mode=block
            Cache-Control:
              content:
                text/plain:
                  schema:
                    type: string
                    example: no-cache, no-store, max-age=0, must-revalidate
                  example: no-cache, no-store, max-age=0, must-revalidate
            Pragma:
              content:
                text/plain:
                  schema:
                    type: string
                    example: no-cache
                  example: no-cache
            Expires:
              content:
                text/plain:
                  schema:
                    type: string
                    example: '0'
                  example: '0'
            X-Frame-Options:
              content:
                text/plain:
                  schema:
                    type: string
                    example: DENY
                  example: DENY
            Content-Length:
              content:
                text/plain:
                  schema:
                    type: string
                    example: '110'
                  example: '110'
            Date:
              content:
                text/plain:
                  schema:
                    type: string
                    example: Sat, 27 Nov 2021 12:38:11 GMT
                  example: Sat, 27 Nov 2021 12:38:11 GMT
            Keep-Alive:
              content:
                text/plain:
                  schema:
                    type: string
                    example: timeout=60
                  example: timeout=60
            Connection:
              content:
                text/plain:
                  schema:
                    type: string
                    example: keep-alive
                  example: keep-alive
          content:
            application/json;charset=UTF-8:
              schema:
                $ref: '#/components/schemas/DeleteStock'
              example:
                path: /api/stocks/1
                timestamp: 2021-11-27 13:38:11
                response_state: true
                status: 200
                data: Deleted
        '404':
          description: Not Found
          headers:
            Vary:
              content:
                text/plain:
                  schema:
                    type: string
                    example: Origin
                  example: Origin
            X-Content-Type-Options:
              content:
                text/plain:
                  schema:
                    type: string
                    example: nosniff
                  example: nosniff
            X-XSS-Protection:
              content:
                text/plain:
                  schema:
                    type: string
                    example: 1; mode=block
                  example: 1; mode=block
            Cache-Control:
              content:
                text/plain:
                  schema:
                    type: string
                    example: no-cache, no-store, max-age=0, must-revalidate
                  example: no-cache, no-store, max-age=0, must-revalidate
            Pragma:
              content:
                text/plain:
                  schema:
                    type: string
                    example: no-cache
                  example: no-cache
            Expires:
              content:
                text/plain:
                  schema:
                    type: string
                    example: '0'
                  example: '0'
            X-Frame-Options:
              content:
                text/plain:
                  schema:
                    type: string
                    example: DENY
                  example: DENY
            Content-Length:
              content:
                text/plain:
                  schema:
                    type: string
                    example: '151'
                  example: '151'
            Date:
              content:
                text/plain:
                  schema:
                    type: string
                    example: Sat, 27 Nov 2021 12:38:35 GMT
                  example: Sat, 27 Nov 2021 12:38:35 GMT
            Keep-Alive:
              content:
                text/plain:
                  schema:
                    type: string
                    example: timeout=60
                  example: timeout=60
            Connection:
              content:
                text/plain:
                  schema:
                    type: string
                    example: keep-alive
                  example: keep-alive
          content:
            application/json;charset=UTF-8:
              schema:
                $ref: '#/components/schemas/StockNotFoundErr1'
              example:
                path: /api/stocks/1
                timestamp: 2021-11-27 13:38:35
                response_state: false
                status: 404
                data:
                  message: Stock not found
                  error: Not Found
      deprecated: false
components:
  schemas:
    GetStocks:
      title: GetStocks
      required:
      - path
      - timestamp
      - response_state
      - status
      - data
      type: object
      properties:
        path:
          type: string
        timestamp:
          type: string
        response_state:
          type: boolean
        status:
          type: number
        data:
          type: array
          items:
            $ref: '#/components/schemas/Datum'
          description: ''
    Datum:
      title: Datum
      required:
      - name
      - amount
      type: object
      properties:
        name:
          type: string
        amount:
          $ref: '#/components/schemas/Amount'
    Amount:
      title: Amount
      required:
      - currentPrice
      type: object
      properties:
        currentPrice:
          type: string
    Data:
      title: Data
      required:
      - message
      - error
      type: object
      properties:
        message:
          type: string
        error:
          type: string
    StockNotFoundErr1:
      title: StockNotFoundErr1
      required:
      - path
      - timestamp
      - response_state
      - status
      - data
      type: object
      properties:
        path:
          type: string
        timestamp:
          type: string
        response_state:
          type: boolean
        status:
          type: number
        data:
          $ref: '#/components/schemas/Data'
    GetStock:
      title: GetStock
      required:
      - path
      - timestamp
      - response_state
      - status
      - data
      type: object
      properties:
        path:
          type: string
        timestamp:
          type: string
        response_state:
          type: boolean
        status:
          type: number
        data:
          $ref: '#/components/schemas/Data2'
    Data2:
      title: Data2
      required:
      - name
      - amount
      - stockId
      - creationDate
      type: object
      properties:
        name:
          type: string
        amount:
          $ref: '#/components/schemas/Amount'
        stockId:
          type: number
        creationDate:
          type: string
    CreateStockRequest:
      title: CreateStockRequest
      required:
      - name
      - amount
      type: object
      properties:
        name:
          type: string
        amount:
          $ref: '#/components/schemas/Amount2'
    Amount2:
      title: Amount2
      required:
      - currentPrice
      type: object
      properties:
        currentPrice:
          type: number
    Data3:
      title: Data3
      required:
      - message
      - error
      - errors
      type: object
      properties:
        message:
          type: string
        error:
          type: string
        errors:
          $ref: '#/components/schemas/Errors'
    Errors:
      title: Errors
      required:
      - name
      type: object
      properties:
        name:
          type: string
    ValidationErr1:
      title: ValidationErr1
      required:
      - path
      - timestamp
      - response_state
      - status
      - data
      type: object
      properties:
        path:
          type: string
        timestamp:
          type: string
        response_state:
          type: boolean
        status:
          type: number
        data:
          $ref: '#/components/schemas/Data3'
    CreateStock:
      title: CreateStock
      required:
      - path
      - timestamp
      - response_state
      - status
      - data
      type: object
      properties:
        path:
          type: string
        timestamp:
          type: string
        response_state:
          type: boolean
        status:
          type: number
        data:
          type: string
    UpdateStockRequest:
      title: UpdateStockRequest
      required:
      - name
      - amount
      type: object
      properties:
        name:
          type: string
        amount:
          $ref: '#/components/schemas/Amount2'
    UpdateStock:
      title: UpdateStock
      required:
      - path
      - timestamp
      - response_state
      - status
      - data
      type: object
      properties:
        path:
          type: string
        timestamp:
          type: string
        response_state:
          type: boolean
        status:
          type: number
        data:
          type: string
    DeleteStock:
      title: DeleteStock
      required:
      - path
      - timestamp
      - response_state
      - status
      - data
      type: object
      properties:
        path:
          type: string
        timestamp:
          type: string
        response_state:
          type: boolean
        status:
          type: number
        data:
          type: string
security:
- {}
tags:
- name: Misc
  description: ''